pcre2grep: correctly handle multiple passes
When the -m option is used, pcre2grep is meant to exit, while leaving
the stream pinned to the next position after the last valid match.

Unfortunately, it wasn't tracking correctly the beginning of the stream
on subsequent passes, and therefore it will fail to use the right seek
value.

Peek the position of the stream at the beginning and while at it, make
sure that the stream passed hasn't been consumed already.