jit: consolidate checks for WX restrictions and make it public
When using the default executable allocator from sljit, the ability
to create maps that are both writable and executable is required,
but some systems had added runtime restrictions to that in the name
of security.

A runtime check was added, but implemented in multiple places so
consolidate the code into a single function, as part of that module
and make it public, so it could be also used proactively and be
consistent with the value returned by pcre2_config to indicate JIT
support is available.

As a sample of how that new information could be useful, change
`pcre2test -C` to indicate if JIT support was disabled at runtime
by using the new API.