jit: explode chars parameter in aarch64 simd
the original implementation abused SLJIT to call a function with
5 parameters, and to avoid using more, packed all chars needed in
a single 64bit variable (really word sized but the implementation
was only available in systems with 64bit words).

this worked for the 8 bit library, but didn't for the others
unless the characters provided to look for where ASCII.

the abuse works because aarch64 ABI requires the first 8 parameters
to be passed in registers, and the registers assigned for it happen
to be the same ones that are available with SLJIT_R[1-8].

instead of trying to pack the characters in a fifth parameter, pass
each character indepedently as 4 additional parameters and fix the
other incorrect assumptions in the code to handle characters wider
than 8 bits.