pcre2_dftables: add some sanity checks an -f option to override
Prevent generating a source file when no LC_ALL environment exist
but allow the user to force it with -f, if needed.

While at it, make sure that the use of UTF-8 locales is flagged
as not safe and use a locale that wouldn't use an UTF-8 charmap
for the documentation.