pcre2posix: protect pcre2_regexec() against truncation
This function was left behind after the migration to PCRE2 with
offsets that are smaller than PCRE2_SIZE, so instead of silently
truncating values it can't handle return an error if the subject
is too big.

Additionally, add validation for the offsets provided by the use
of REG_STARTEND.

While at it, use casts for regoff_t instead of int, for clarity
and to make it easier to update if regoff_t is ever changed.