pcre2_compile: tighten recent changes
The code intended to error for numbers that were bigger than
MAX_GROUP_NUMBER so change the previous max value to match.

Document the previous limit with an assert, as it is required
to avoid integer overflow and while at it change the variable
used to reference the character being checked for consistency.