reimplement asserts with better focus in debug mode
The original asserts weren't very useful in debug mode as they
were lacking information on where they were being triggered and
were also unreliable and dangerous as they could result in
important code being removed and trigger crashes.

Instead of implementing one generic assert for both modes, build
a more useful one for each mode, but to make sure that the non
production paths are not being unnecessarily eliminated, allow
for a parameter that could be used to indicate if that functionality
is desired or not.

Reinstate all original assertions to use that instead, and make
sure to set the parameter to a safe value in the one that is known
to cause problems with the previous code.