ci: upgrade build linux job to latest Ubuntu
Most of the uses of `PCRE2_UNREACHABLE(0)` are at the end of `case`
and therefore in non debug builds, the assertion "should" tell
the compiler that a "fall back" is not possible, but the version
of gcc used in Ubuntu 22.04 has a bug and will instead see the
assertion as additional code that doesn't have a `break` after
and therefore trigger `-Wimplicit-fallthrough` warnings instead.

Update the job to use Ubuntu 24.04 that provides gcc 13.2 and
that doesn't have the bug anymore, and while at it update all
jobs to error on warnings so that failures will be more visible.