pcre2_error: add runtime overflow checks to pcre2_get_error_message()
As documented[1] in the API, error messages are meant to be short enough
to fit in a small buffer, but there are no checks to ensure that its
maximum expected size is enforced, so add an imperfect one.

Extract the logic to count messages into a helper function, and take the
opportunity to improve the code and make it almost twice as fast.

Add a test showing the larger message and an assertion so a violation
could he hopefully caught at runtime (with a debug build).

Whike at it, make the use of CHAR_NUL to represent the terminating C
string, consistent.

[1] https://www.pcre.org/current/doc/html/pcre2api.html#geterrormessage