pcre2_error: protect pcre2_get_error_message() and its users
When an error code absolute number is too large, the buffer will
be left mistakenly uninitialized. Additionally, if INT_MIN was
given, the implementation will trigger undefined behaviour in two
complement systems.

Make sure to always zero terminate the buffer when returning
PCRE2_ERROR_BADDATA by centralizing all callers through a helper,
and skip finding a non existant message with INT_MIN.

Add functionality to pcre2test to test this issue and while at it,
change the type for errorbuffer so it is aligned for all libraries.