pcre2test: add support to unit test error codes
Allow calling `-error` without a provided error code list to loop
through all error codes looking for problems and add it to RunTest.

It is intentionally not doumented, as it will only work in debug
builds and needs additional system dependencies that might not be
relevant or available to non debug pcre2test users.

Includes minor fixes to the current `-error` handler.