pcre2test: refactor memory debug support with newer C
Move reporting for malloc/free helpers to stderr so it doesn't
mix up with standard output,

Add support for testing a not NULL data pointer in the memory
management functions, and groups most variables related to this
testing in a single global struct for easier management.

It is the first user of Designated Initializers from C99, and
instructions are added for any compiler that might not support
that, but since we had required a C99 compiler for a while it
is unlikely to be an issue; either way, it is better to know now
when it can be easily worked around than later.

While at it, avoid truncating length of strings in non 8-bit mode.