jit: generalize off by one fix from #86
The ARM64 SIMD implementation use their own C based function
generator and compiler intrinsics, unlike the other architectures
and therefore has known issues that are yet to be fixed.

A crasher was reported and fixed for 10.42, but there was at the
time difficulty reproducing it, so a minimum fix was committed.

Generalize that logic for all other generated functions to fix
similar issues on them.