jit: prevent subject buffer over reads in arm64 SIMD helper
prevent CWE-126 from the fcs_* generated functions for NEON and
that could result in crashes (observed in OpenBSD arm64).

refactor fix from #86 to better reflect expected logic by returning
early with a match failure also when the initial call provides a
subject pointer at or after the end of the subject.